lift and shift inline functions to  files
